% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patviewer.r
\name{patview.download.files.list}
\alias{patview.download.files.list}
\title{Downloads list of files from PatentsView.org that are available for download}
\usage{
patview.download.files.list(patview.files.list.path = file.path(getwd(),
  "patview.files.list.html"))
}
\arguments{
\item{patview.files.list.path}{A path to the html page with the table that contains the list of files to download. Will try to download it if it not founl. The default is in R working directury "patview.files.list.html"}
}
\value{
Saved file path.
}
\description{
Downloads list of files from PatentsView.org that are available for download
}
