% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patviewer.r
\name{patview.filter}
\alias{patview.filter}
\title{Filter tables of PatentsView.org bulk data}
\usage{
patview.filter(file.pattern, ..., file.dir = file.path(getwd(),
  "patview-rds"), pbapply = TRUE, cols = character(0))
}
\arguments{
\item{file.pattern}{A pattern for getting a file or a set of files (data batches)}

\item{...}{A filtering conditions to fetch certain rows. (See dplyr::filter)}

\item{file.dir}{A path to directory with .rds files. Default is in working directory "patview-rds"}
}
\value{
A data.table with a subset the data.
}
\description{
Similar to dplyr::filter but for tables of PatentsView.org bulk data saved in multiple .rds files
}
