% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patviewer.r
\name{patview.full.deploy}
\alias{patview.full.deploy}
\title{Deploys PatentsView dataset. Needs lot of disk space and memory}
\usage{
patview.full.deploy()
}
\value{
files path
}
\description{
Deploys PatentsView dataset. Needs lot of disk space and memory
}
