% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patveiwer.r
\name{patview.get.files.list}
\alias{patview.get.files.list}
\title{Gets list of files from PatentsView.org that are available for download}
\usage{
patview.get.files.list(patview.files.list.path = file.path(getwd(),
  "patview.files.list.html"))
}
\arguments{
\item{patview.files.list.path}{A path to the html page with the table that contains the list of files to download. Will try to download it if it not founl. The default is in R working directury "patview.files.list.html"}
}
\value{
Saved file path.
}
\description{
Gets list of files from PatentsView.org that are available for download
}
