% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patveiwer.r
\name{patview.save.rds}
\alias{patview.save.rds}
\title{Reads PatentsView .tsv files and saves them to .rds (also optional in batches)}
\usage{
patview.save.rds(file, dir = file.path(getwd(), "patview-data-tsv"),
  dir.rds = file.path(getwd(), "patview-data-rds"), batch.lines = 0,
  file.lines = 0)
}
\arguments{
\item{file}{File name (.tsv is expected)}

\item{dir}{Directory where the file is. Default is in working directory "patview-data-tsv"}

\item{dir.rds}{Where to save. Default is in working directory "patview-data-rds"}

\item{batch.lines}{How many lines is to read in one batch (10^7 is recomended). The default is 0, meanning that reading will be done in one batch as a single file.}

\item{file.lines}{Length of the .tsv file. The default is 0. If is is not changed and batch.lines is specified then it will try to calculate it with grep.}
}
\value{
Saved file(s) path.
}
\description{
Reads PatentsView .tsv files and saves them to .rds (also optional in batches)
}
