% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patveiwer.r
\name{patview.unzip.file}
\alias{patview.unzip.file}
\title{Unzips (PatentsView.org bulk data) files.}
\usage{
patview.unzip.file(zipfile, zipdir = file.path(getwd(), "patview-data-zip"),
  exdir = file.path(getwd(), "patview-data-tsv"))
}
\arguments{
\item{zipfile}{File name to unzip.}

\item{zipdir}{Default is in working directory "patview-data-zip"}

\item{exdir}{A path where to extract zip file. Default is in working directory "patview-data-tsv"}
}
\value{
Unzipped file path.
}
\description{
It also renames the files ("_" -> ".")
}
